Yet another check removed from postHandler of accept proposals list.
I believe the purpose of that check was to make sure the proposal was not accepted first and then the project was withdrawn. It was not correct:
- a student teoretically might have another project for the same organization
- it was not executed within a transaction

It is also not necessary, as acceptProposal function will take care of all situations in which the proposal was already transformed into a project.