convert_profile job generalized to handle all program scoped models.
For each kind of such models, scope property has to be copied to a new program property. It is better to use the same code than to create a new module with the very same code for each model.

Another important change in this commit is that scope is actually not accessed via RPC to datastore. Instead, its key is obtained by get_value_from_datastore. This is crutial here, for no cross entity group transactions are needed right now.

Previously, calling entity.scope would actually accessed Program entity, so another entity group (the one associated with the program) was involved in the transaction. It was a big problem, because there is only a few programs. Parallel jobs accessed the same entity at the same time. When the transaction was supposed to be committed, it failed, as optimistic concurrency counters did not match. Transactions retried, but the situation happened again.

This actually answered one dobut I had about XG transactions. It was not clear to me whether readonly operations on one entity group in another transaction could lead to failure of the given transaction (if it also need readonly access). Apparently, it does.