ConvertSponsorScopedModel map-reduce script added.
It is similar to recently added ConvertProgramScopedModel script, but scope property is copied to 'sponsor' property insead of 'program'. Probably the previous script could have been adapted to accept destination field name as a parameter, but it is not a problem to add a new one for this purpose.

Also, the GSoC prefix is removed from the name of the old script. It is more appropriate, as usage is not limited to GSoC module.