model_pacifier MapReduce script added.
Sometimes some old, unused and deprecated models are removed. Anyhow, the entities are still stored in the datastore, waste the disk space and have to be manually removed. This script will take care of that.

It accepts model name (fully qualified with the module path) as its parameter which makes it more versitile.