RichBool class added.
Its main purpose is to hold a bool value with optional extra argument which may be passed around and read by interested clients. Much more details are included in the doc string.

Typical usage of that class is if a caller of a function which would normally return a bool is interested in getting more details. For example, why a specific value was returned?