Validators for database properties.
At the moment of this revision they are not needed. AppEngine database models may use EmailProperty and LinkProperty to enforce input validation. Anyhow, the plan is to update at least the most important (and most frequently accessed) models to take advantage of NDB API.

In the new approach, all those specific properties are dropped. Only a genenric StringProperty is left. Fortunately, an option to specify a validator for each field was added. Therefore, correct values of these two deprecated properties may be enforced by the newly added validators.