Replace outdated module reference and test changes.
The soc.logic.exceptions module was removed in revision
f33721c323db, so this change utilizes the new exceptions as
defined in melange.request.exception. Also makes minor fixes to
as test where the set function is called with multiple arguments
instead of a list and accidental method call instead of a
string reference.