Survey type concept added to ProjectSurvey.
So this is model class which stores surveys that are to be taken by students who
participate in the program. Theoretically, it should support as many surveys per
program as possible. Practically, there have always been only two surveys and
only those are currently supported.
It is better to have these strings, which will work from now on as their types,
here than wandering around the codebase. Also, these strings are important, as
they create a part of key name for ProjectSurvey model.