urls module added to summerofcode helper module for views.
Right now, it contains UrlNames class with URL names used for the views in this module. It is a class so that it can be injected directly to request handlers. It will make it easier to handle URL managing.

Currently, we have url_names module which contains raw string constants. The problem is that these modules need to imported directly by views. Sometimes, semantic difference between module specific views (i.e. gsoc or gci) is that they need to access different URL names, which are also module specific. Because of that, the code is outlined in the following way:

class GenericView():
  def foo():
    # do some stuff
    url = self._getUrlName()

  def _getUrlName():
    raise NotImplemented

class GSoCSpecificView(GenericView):
  def _getUrlName():
    return 'some specific url'

class GCISpecificView(GenericView):
  def _getUrlName():
    return 'some specific url'

The problem is that these specific views are needed only to override the abstract method. If these dependencies can be injected to GenericClass, these extra classes are not needed anymore.

This view will probably be extended with some extra features. It still has to be figured out what the best for project is.