CSS and HTML for displaying "You are logged in as ..." message at the top of page.
Currently there is no one clear place at which a message like that is located. It is sometimes at the top of some forms, but it is not always there. There should be one well-defined place where a user may quickly find out what account he or she is logged in with.