A number of access checking tests for Take Org Application page.
There are different scenarios of who and when is trying to access the page.

Currently, one of the tests - testSummerOfCodeProfileAccessDenied - fails.
In order to be able to access the page, a user should have AT LEAST one mentor or org admin role in previous GSoC or GCI programs. Unfortunately, right now it is sufficient for a user to have a non-student profile. So, for example, if an user registers for GSoC as a mentor, but is not accepted by any organizations, he or she will still be able to take org application.

It will be fixed in the next commit.