Changes "Preview Mentor Evaluation" page to prevent issue 1873.
The issue above is related to "Preview Student Evaluation" page, but it could be obviosuly reproduced here. In this revision raise_not_found is set to True, as it makes no sense to preview non-existing evaluation.

This change is analogous to revision e62161f that fixed the issue for student evaluations.