Only instances of Template class are rendered in assertRenderAll.
This is not a problem right now, but some tests will fail upon upgrade to Django 1.4. I have not figured out how life of a django template context looks like, but, for some reason, returned context dictionary is extended with 'base' item. The associated value is of BlockNode class.

Instances of that class also have render() function. Unfortunately, that function require another context attrubute:

def render(self, context)

In assertRenderAll function, it is only checked if the given object has 'render' attribute. It is not sufficient, because it assumes that it is a function with no extra parameters. Therefore, the following error is returned.

TypeError: render() takes exactly 2 arguments (1 given)

This change will add an explicit check if the tested object is an instance of Template class. Objects of that class have render() function and those are the objects that are really supposed to be tested by this function.