There are module specific UserReadOnlyTemplate subclasses.
There was a tircky and difficult bug to find. The problem was that each Template require its own template_path that will be used to render actual HTML code with the specified context. Generic UserReadOnlyTemplate class was used by both GSoC and GCI, but that class never declared any template path.

If someone goes to actual "Show Profile" page, it will be noticed that the section concerning User data, i.e. username and account, will not be displayed.

This is because ModelReadOnlyTemplate, from which UserReadOnlyTemplate class inherits, has render() function that refers to self.template_path. The problem is that Django's functions fail silently. loader.render_to_string does not raise any actual exceptions. The only sign was the absence of user data in actual view.

The tests did not fail either. The response code was OK and the expected HTML template (''modules/gsoc/_readonly_template.html') was used for another Django template.

I found out this error when I upgraded Django to 1.4 locally.