Middleware class to handle file uploads added to test_utils module.
Currently, it is not really possible to test file uploads. The reason is that the regular blobstore middleware class handles only content with very specific content. For example, they need to have particular headers and contain blob-key.

The problem is that this is normally added by AppEngine at some point. Unfortunately, this step was not executed by the test framework.

Actual file content will not be available to the views. This class just creates some mock data so that the views will be able to tell if a file was supplied or not.