schemaIterator and fieldsIterator function added to SurveyRecordReadOnlyTemplate class.
Their purpose is to replace the removed __iter__ function. The first one will iterate through all the fields of the survey schema without going through any model fields. This is what currently is needed for mentor and student evaluations. The second function will iterate through the declared model field. It is useful for org application surveys.