Small fixes to mistakes that broke tests.
I added a call to redirect.program() so that it clears the RedirectHelper's args and kwargs then sets the program and sponsor in its kwargs, which are used when it creates the gci_conversations url. When testing, data.profile is sometimes None, so the code just before mine doesn't create the dashboard url, thus the RedirectHelper has uncleared args and lacks the needed kwargs. Calling redirect.program() is what I should have done from the start.

I also changed my condition from checking for data.gae_user (which was misguided) to check instead for data.user, which is the attribute I am actually using.