profile property moved up to soc's RequestHandler class.
The implementations for GSoC and GCI were both identical except for the models class that was used. The code was, therefore, duplicated. Also, profile() property could not be accessed from soc-level code (for example tests). It is beneficial just to "inject" good model.

This solution is not pretty and perfect, but it is still improvement.