userFromKwargs removed from Mutator class.
Instaed, it url_user, which was actually set by that function, is a property of RequestData object.

It is better, as Mutator instance does not have to be involved in accessing the user. Mutator class is also a bit unsafe and unclear. It adds some arbitrary fields to RequestData object.

Also, unit tests are added in this revision. It is also a bit more flexible to test RequestData directly. Mutator class was not covered by tests at all, but its methods would have to be called and then side effects on the RequestData object would have to be checked.