connection_url item is not generated.
The code is commented out and TODO is added. The problem is that the function uses data.redirect object in
order to obtain URL. However, show_user_connection is not defined for GCI. It could be definied, but on the other hand, RedirectHelper (as a very stateful object) is going to be deprecated. Linker should be preferred. Indeed, Linker WILL be used here at some point.