URL pattern of ManageConnectionAsUser change to USER_ID.
This is because connection entities have IDs instead of key names. If the pattern is

<sponsor_id>/<program_id>/<user_id>/<connection_id>

the connection entity can be retrived by db.get_by_id() function. In the previous approach, the pattern was:

<sponsor_id>/<program_id>/<user_id>/<org_id>

In order to fetch connection, db.Query had to be used (filtering by org and user).