IsUrlUserAccessChecker class added.
It checks whether the currently logged in user is the one who is defined in the URL data. For example, the following URL pattern:
<sponsor_id>/<program_id>/<user_id>
may be accessible only to users for whom <user_id> corresponds to their own username. In other words, a user, whose name is 'foo', can only access google/gsoc2013/foo.

Also, unit tests are included in this revision.