Merge conflicted files from master.
The goal of this branch is to completely remove the existing
connection functionality and replace it with the new implementation
from drew_connection_rebuild since the latter's history is too
messy for an outright commit.

These five files contained commits from both that branch and master
and needed to be merged, which I did by checking out each file from
drew_connection_rebuild with the -p (patch) option and choosing which
chunks to checkout. In a couple of cases Git treated different changes
as one hunk so minor changes from master needed to be manually
added back in in favor of getting the new functionality added.