createBlob added to SoCTestCase class.
It is a utility function that saves a blob for the specified file name. Actually, it should be much simpler than that: createBlob function should return BlobKey of the newly created blob. For some, unknown to me, reason it returns generic Key instance. Unfortunately, some AppEngine model properties require BlobKey instances or raise an error. This is why this function creates a blob and then generates explicit BlobKey for the same file name.