ConnectionForm to start connection as user simplified.
In particular, no user_role is displayed. It was decided that the user would not have option to pick whether he or she wants to be org admin or just a mentor. The user would only tell that he or she is interested in having a role with the specified organization. If the user starts a connection, there is actually no choice at ll: it means the person wants a role. It would be pointless to start a connection to say "I do not want a role with you guys."