ReadOnlyTemplate class added to GCI templates.
It will be used to display arbitrary items as tables. It is a bit similar to classes provided by readonly_template with the exception that it is not bound to any db.Model. It is much easier to use this class for items that do not reflect model properties.

In the future, it will probably be moved to Melange, as the only Code In-specific element is template_path.

Also, simple unit tests are added.