readonly.less is a part of bigger style.less template.
Specifically, the latter CSS template imports the former one. In the result, eventual style.css stylesheet contains all CSS rules defined in readonly.less.

Because of this change, readonly.less can refer to variables defined in style.less. Of course, size of the generated style.css file will be bigger but it does not really matter, as "readonly" style is very short.