Positional arguments removed from all model forms.
Instead, all arguments are passed as kwargs. The only exception for now is bound_field_class which still has to be at position one. It will be changed at some point in the future.

The point is that the base ModelForm class does not use any positional arguments. For some reason, we usually put data in a positinal manner, like:
form = Form(data)

It worked only because data is the first kwargs arguement of ModelForm initializer. Anyway, it was difficult to add new kwargs (optional) arguments to new form implementations, as kwargs and positional arguments might get easily messed up.