addFilterToQuery function added.
It is a typical utility function that adds a filter to the specified query. It is simple but callers do not need to distinguish between lists/tuples and single objects. It is the polymorphism that AppEngine achieves through adding operator to the first argument of filter function.

This approach was already used in _handleExtraAttrs function. It is moved to appengine.db so as to make it accessible to all modules.