Don't develop YAML lib of Google App Engine.
We have no idea why this is in place and it appears that removing it
doesn't break anything.

This removal is now necessary because the "develop" part of buildout
is run at the beginning, and, since we're going to install the App Engine
SDK during build, the directory this line is trying to access is not
present when the attempt is done.