Changes to how universities are uploaded.
The previous approach to allow as many universities per one function call as possible was too optimistic. The first problem is that dynamic calculation of entity's size is slow and takes a while especially for larger structures. The very next attempt was to simply perform that check after certain steps only, for example every 1000 iterations. It did not help at all. The next simplification was an explicit declaration of how many universities per entity may be stored. In that case, nbd.pull_multi was not able to handle 5 large entities and exceptions were raised.

This revision simplifies the process by allowing to create one cluster up to 5000 entities at a time. It is simple but acceptable, as only program administartors will upload that list once a year.

Also, unit tests are adjusted.