Concept of universities renamed to schools.
This is because 'university' string is longer than 'school'. Also, all universities are schools but not all schools are universities. There are colleges and polytechnics. Last but not least, 'school' name is more appropriate if used for Google Code In.

At this point, tests are broken, as imports do not work. They will be fixed in the upcoming revision.