Memcache is used to access schools.
When a list of items is retrived from Blobstore for the first time, the client caches the final structure (as JSON string) in memcache. Thus, next clients may access it much faster.

Please note that we do not need to care about thread safety here. Setting the item in memcache does not need to be synchronized. In the worst case, a few clients will get the value from Blobstore and then put it to memcache, but we do not need to worry about it.

Also, a defaultdict is used instead of a regular dict.

Last but not least, a unit test that shows that school data is really retrieved from memcache is added in this revision.