When a connection is created within a transaction and it turnes out that there is already one, BadRequest is raised.
It is more appropriate, as the resource itself is not forbidden. It is just not possible to handle the data, because state is not legal. Anyway, it is very unlikely that it will ever happen. The prerequisites are first checked by the access checker. The state would have to change between that stage and the moment when transaction begins.