Organization actions on connections are handled now.
What is not supported right now is that when it is not possible to complete an action for some reason, the reason is not propagated to the end user who has initiated the action.

A number of unit tests are also added. Their purpose is to cover all state transitions.