predefined_schools_counter field removed from Program model.
It is not needed anymore, as schools are stored in a blob right now. The main purpose of this field resulted from imperfection of storing large number of schools in datastore. It was not possible to upload and put all schools in a single request. Therefore, the global counter was added for user's convenience: if the number get higher, the last batch was uploaded correctly.