ConnectionList and UserConnectionList templates added.
These are two abstract classes. The first one is more general and defines general behavior of a template to list Connection entities. UserConnectionList is based on user's perspective and lists all connections of the specified user.

Some things might be implemented better here, the current approach to lists is used with no improvements or modifications. It will be changed when new module that supports lists is production ready.