ListConnectionsForUser page added.
It is simply a page that lists all connections for a user who is specified in the URL. As a part of this functionality, CIUserConnectionList class was added. It extends and implements all methods of UserConnectionList by specifying url_names and templatePath. HTML templates are also added in this revision.

Last but not least, a couple of simple unit tests are added. One of the cons of using the current lists is limited testability. The only thing that can be done is checking that number of elements in the list is correct.