ListConnectionsForOrgAdmin page added.
It lists connections that can be managed the visiting organization administrator. As a part of this revision, CIOrgAdminConnectionList and OrgAdminConnectionList classes are added. The latter class is abstract and provides methods to list particular connection entities. The former one is a concrete implementation that defines url_names and templatePath.

Also, a few simple unit tests are added.