GCI Dahshboard contains Connection related items now, whilst Invite based ones are removed.
One thing that is missing right now is an item that allows Org Admins to start connection with other users. There was a corresponding field regarding Invitations. Anyway, user experience was not satisfactory and it will be improved before such an item is added here.

When a user clicked on "Invite" item, he or she was taken to a list of organizations and the poor person was required to pick the right organization first. That would make a lot of sense if an average joe admins, let us say, 10 organizations. Practically, it has not happened ever. Suffice it to say that GCI accepts ~10 orgs in total, but it is also true for GSoC. In large majority of cases, one user admins only one organization. Therefore, he is shownd a list with one row that he has to click.

It should be one step operation.