seedUser function added to profile utils.
This function will be used to seed User entities. It is not trivial and standard seeder logic cannot handle it properly, unfortunately. This is mainly related to account property whose type is UserProperty. The underlying type is google.appengine.api.users.User and it is based on environmental variables.

Each AppEngine's User has user_id property and it is actually used by Melange to retrieve entities associated with logged in users. It is actually duplicated and stored in Melange's User model. This is the first reason why regular seeder cannot handle it properly: it just finds user_id a string property and simply generates a random string.

Another reason is that, for some reason, account's user_id is not set after the entity is initially put in the datastore. In other words:
user = user_model.User()
user.put()
print user.account.user_id()
will print None. When the entity is fetched from the datastore, it prints out some reasonable value. The new function does exactly that.

Also, this function is not in the scope of ProfileHelper. This class has a lot of states (for example program) and other values that have to be set. It makes it relatively hard to use. Personally, I think that helper seeding function should be as simple and as versatile as possible. The methods of ProfileHelper class, however, do call the new function.