NonStudentAccessChecker replaced by NonStudentUrlProfileAccessChecker.
Currently, it is used by StartConnectionAsUser view. The URL pattern is as follows:

/connection/start/user/<sponsor_id>/<program_id>/<user_id>/<org_id>

It is <user_id> user, defined in the URL, who must not have a student profile for the program. Technically, someone else may be allowed to access the page as well. Then even if data.profile.is_student is False it does not mean that data.url_profile.is_student is also False. It does not really matter in this case, as it is also checked whether data.profile == data.url_profile.

Also, unit tests are modified so as to fit the new name. In testStudentAccessDenied, data.profile is also seeded and is_student is explicitly set to False.