A few minor changes to NonStudentUrlProfileAccessChecker.
Mainly, a message that is raised when url profile is stored by a variable with a slightly different name.

Also, real entities and request kwargs are used by the tests.

Last but not least, if not data.url_profile check is removed. This instruction itself will throw a NotFound error now.