Key column of MyEvaluationsComponent and StudentEvaluationComponent are added manually.
This is the way how it was constructed before revision 9a07af6d8059. That revision extended addKeyColumn function so that it was able to generate different strings for entities with parents.

Unfortunately, that change does not work for the list of evaluations. It is a very complex class/list: it is supplied all evaluations for the program in the constructor. Then, student project is retrived by a query defined in getListData and a custom row_adder function is supplied. This function is responsible for binding evaluations with the project and for adding a row for each evaluation. Complicated, isn't it?

Anyway, the problem is that lambda expressions used in that class are passed the following two arguments:
- project entity
- evaluation entity

addKeyColumn function is based only on the first argument, whilst it is constant for all evaluations. Therefore, the keys are not unique.

It fixes an error that students and mentors are not able to access their final evaluations.