UserSettings model added.
It will be used to specify site-wide settings by users. The first setting that is declared here as of right now is view_as property which is a reference to another user. If set, the user will pretend to be logged as that other user. It is not supported yet.

Therefore, developers will be able to easily see content of pages in context of other users. It will help to debug and reproduce problems which are reported by others.