Initial version of Landing Page is added.
It will be changed somehow in the future, as GSoC and GCI modules should not be accessed so explicitly by the main module.

HTML templates are not complete at this point. They will be and CSS will be added when the final version is delivered by EchoDitto.

Also, a few unit tests are added.