SoC's base.html template renamed to root.html and moved to Melange package.
The reason is that base.html for GSoC and GCI is one level lower: it defines the main layout and leaves a developer with a task to provide page content. base.html for SoC just defines base HTML structure up to body element. It has been renamed, as there will be base.html with layout definition for Melange package (i.e. related to neither GSoC nor GCI) and there would be name collision.