Add new fields for AnonymousConnections.
The existing implementation will no longer work in light of the
relatively recent host of changes to Connection. Instead, when an
org attempts to invite an unregistered user they will receive an
email notification with this object's token in it that can then
be used to enroll as a mentor or org admin for the org.