"Contact Us" section added to Landing Page.
It will list a list of lists/icons that allow users to get in touch with the authors/maintainers of the site. It will be displayed only when there is at least one channel defined in site settings.

At this moment, the CSS does not look very well if not all channels are set. It has to be adjusted.