Add tests, query helper method and return connection entity instead of key.
This query is performed in mutliple places and I encapsulated it
within a method for obvious reasons. Tests have been added for this
method and also test to verify that an AnonymousConnection is
deleted after it is activated since there aren't really any
more apropriate places to do it (should be a cohesive process).