Update redirect method with newer AnonymousConnection fields.
I had previously planned on just adding this same basic functionality
to ShowConnectionForUserPage, however I didn't realize Daniel had
implemented this. So I updated the method to expect a token instead
of a hash along with the docstring and a user will now be redirected
to the page where they can activate an AnonymousConnection.