RequestData's user property takes view_as property into consideration.
If a user is a developer, he or she may have 'view_as' property set in their UserSettings entity. Therefore, if the property is set and the user exists, this underlying user will be returned as the value of property. Thus, the developer will be able to see the page from that person's perspective.

It is not likely, but if the underlying user does not exist, BadRequest will be raised.

Also, unit tests for user property are added in this revision.