_updateSeenByProperties function added to connection logic.
It updates seen_by_org and seen_by_user properties based on the side which has initiated some action the specified connection. For example, if the action is taken by the organization, seen_by_org is set to True, because someone from the organization has already seen the effect, i.e. the organization admin who takes the action. On the contrary, seen_by_user is set to False, because the user has not visited Manage Connection page.