MessageFromHandler accepts author key as another argument of its constructor.
Therefore, author of the message is not always set to data.url_profile.key() which was not correct. When organization admin open 'Manage Connection As Org' page, the author is specified by data.profile, whereas data.url_profile refers to the connected user.

This change fixes testSendNewMessage test belonging to ManageConnectionAsOrgTest. The test is not thus skipped.