CSS for red theme added for GCI blocks.
In order to become a red block div element must have two classes specified:

<div class="block red-block">
</div>

Just a note.
I think that block class defines too much behavior as of right now. Not only does it specify how the blocks should be outlined, but it also describes how they should look. The problem is that many other classes want to take just the former and specify the latter on their own. I am not sure if it would not be clearer if the blue color was pulled out from "block" class to a separate one too. Anyway, I need to talk to someone who knows CSS better than I do first.