Initial draft of AbsoluteLinker that constructs absolute links for Melange.
It is stateless except for hostname that is immutable, so there is also one instance needed throughout the site.

There is a TODO added so that hostname should not be obtained from datastore. It might be possible to get that information directly from the environment so that no RPCs or, God firbid, transactions are required.