Add email field to AnonymousConnection module.
As discussed in comments on revision 76c3ed1e3e1f, this will allow
us to monitor the correlation between users who register with different
emails than the one used in the connection. This will also allow us to
restrict users from registering with emails other than the one to which
the connection was sent should we choose to for security purposes.

Includes adjustments to all involved modules and tests.