Sponsor and program related kwargs are obtained from program key.
No RPC call is required, as the hardcoded pattern used for program's key name is as follows:
<sponsor_id>/<program_id>

Both names can be easily computed by splitting the whole key name.

Retriving a program entity becomes problematic if the function is executed within a transaction, as it probably triggers a new DB call on a different entity group.