DjangoTestCase class is modified so that it can be used by Django 1.4.
Test suite for Django 1.4 is weird.

First of all, they introduce a new concept of SimpleTestCase class. This should suit Melange, as the purpose is to separate tests cases which use a database backend provided by Django from test cases which do not need to set up any environment useful for data persistence. This is just great, because Melange uses AppEngine backend.

Unfortunately, the world is not so sweet. Django moved some non-database related stuff up to SimpleTestCase, but for some reason they left other functions (which are not associated with DB connection at all!) in TransactionTestCase. Fortunately, they might have realized their mistake and all functions are located in SimpleTestCase at the current HEAD commit. Maybe it will ship with Django 1.6?

Arguably, we could also stick with TestCase or TransactionTestCase for now. There is nothing wrong with doing these initializations as long as they do not imact test behavior. Unfortunately, these tests fail to load. I do not understand why it works that way, but Django assumes that the DB backend is not the dummy one that we use. Most of interface funtions are not supported by dummy and if invoked, they call complain() function and _pre_setup function fails. Personally, I think that unsupported functions should be handled gracefully.

These is the reason why _pre_setup function does not call super and the DjangoTestCase function does not inherit from SimpleTestCase.