getOrgAdmins function added to org application logic.
It is a very important function and its implementation is much more complex than it should be. This is because Melange did not require takers to have a profile some time ago in the past. It is not the case anymore, but I ran a mapreduce today in order to check how many Profile entities are missing and there is a lot.

It is not very important, but at some point int the future I would rather store profile entities.

It is also related to issue 1769.