prefix field added to Program classes.
Hopefully, it is going to be there only temporarily. It is helpful, beucase there are duplicated functions which have the same logic but they just differ so that one of them use 'gsoc_program' and the other 'gci_program'. I just looked at it briefly and it seems unnecessary but it is not a good time to remove it now.