StartConnectionAsUser URL pattern changed from USER_ORG to ORG.
Therefore, no user id will be present in URLs. It is not needed at all, because a user should be able to start a connection only for himself or herself. It was present there, as I generally had this idea to make paths independent of data.user property. The main problem was that it was very difficult from developer's perspective to open a page on someone else's behalf. By adding this user ID, developer could just put it in and see a paricular page as a person with that given ID.

Anyway, in the foreseeable future, there will that view-as thing which is more generic and flexible. Thus, URLs can be simpler.