"connect with us" template redefined.
It now uses the new set of icons which were delivered by EchoDitto. Therefore, it was relatively easy to add a Google Plus icon.

Twitter icon had to be removed, as it is not part of the images they provided. It is not a problem, as Twitter was actually removed from program settings. So was Facebook. If we ever need, we can just ask for new icons. Adding Google Plus is more important.