request_data added to SiteForm.
It is required because one of the fields of SiteForm is Terms of Service which is a reference to Document. GSoCBoundField will pick a helper widget which expected request_data to be set. Why didn't it fail before? I believe that Django just failed silently and did not render that one field.