executeTasks function added to DjangoTestCase.
Because of this change, it will be possible to execute GAE tasks more easily. I cannot go anywhere further at this moment, i.e. to SoCTestCase, because the tasks are executed by post client provided by Django.

It is technically possible to use general WSGI request (not Django), but it would be very difficult, because in our case Django layer also binds URLs with request handlers. Anyway, it is totally fine.