StartConnectionByUserContextProvider and StartConnectionByOrgContextProvider classes added.
These classes offer one public function getContext and are similar to userConnectionContext and orgConnectionContext, respectively. The main difference is that these are classes, so it is more elegant to pass them as function parameters. It is still not exactly what I would really want, as I have in mind something like ContextProvider (or NotificationContextProvder) interface. It is not trivial, as different provider implementations may need different function parameters.